Add files via upload
Two output csv files are included in the weather_csv folder. The weather_df2 is the correct one to look at. The reason that weather_df has some problem is because 2 reasons. Reason 1: First column was not cleaned. Reason 2: The max temp was set as absolute temp instead of Fahrenheit. Anyways, differences could be found when you compare them.